refactor(db): modularize MDX data handling by extracting utility functions (#580)
