+++ apps/web/src/lib/db/portfolio.ts
@@ -1,65 +1,5 @@
-import fs from 'fs';
 import path from 'path';
-
-type Metadata = {
-  title: string;
-  publishedAt: string;
-  summary: string;
-  category?: string;
-  banner: string;
-  alt?: string;
-  image?: string;
-};
-
-function parseFrontmatter(fileContent: string) {
-  let frontmatterRegex = /---\s*([\s\S]*?)\s*---/;
-  let match = frontmatterRegex.exec(fileContent);
-  let frontMatterBlock = match![1];
-  let content = fileContent.replace(frontmatterRegex, '').trim();
-  let frontMatterLines = frontMatterBlock.trim().split('\n');
-  let metadata: Partial<Metadata> = {};
-
-  frontMatterLines.forEach((line) => {
-    let [key, ...valueArr] = line.split(': ');
-    let value = valueArr.join(': ').trim();
-    value = value.replace(/^['"](.*)['"]$/, '$1'); // Remove quotes
-    metadata[key.trim() as keyof Metadata] = value;
-  });
-
-  return { metadata: metadata as Metadata, content };
-}
-
-function getMDXFiles(dir: string) {
-  return fs.readdirSync(dir).filter((file) => path.extname(file) === '.mdx' || path.extname(file) === '.md');
-}
-
-function readMDXFile(filePath: string) {
-  if (!filePath || !fs.existsSync(filePath) || (path.extname(filePath) !== '.mdx' && path.extname(filePath) !== '.md')) {
-    throw new Error('Invalid file path or file does not exist');
-  }
-  let rawContent = fs.readFileSync(filePath, 'utf-8');
-  return parseFrontmatter(rawContent);
-}
-
-function extractTweetIds(content: any) {
-  let tweetMatches = content.match(/<StaticTweet\sid="[0-9]+"\s\/>/g);
-  return tweetMatches?.map((tweet: any) => tweet.match(/[0-9]+/g)[0]) || [];
-}
-
-function getMDXData(dir: string) {
-  let mdxFiles = getMDXFiles(dir);
-  return mdxFiles.map((file) => {
-    let { metadata, content } = readMDXFile(path.join(dir, file));
-    let slug = path.basename(file, path.extname(file));
-    let tweetIds = extractTweetIds(content);
-    return {
-      metadata,
-      slug,
-      tweetIds,
-      content,
-    };
-  });
-}
+import { getMDXData } from '@/lib/db/utils/get-mdx-data';
 
 export async function getPortfolioPosts() {
   return getMDXData(path.join(process.cwd(), 'src/contents/portfolios'));
