+++ apps/web/src/lib/db/utils/get-mdx-files.ts
@@ -1,0 +1,6 @@
+import fs from 'fs';
+import path from 'path';
+
+export function getMDXFiles(dir: string) {
+  return fs.readdirSync(dir).filter((file) => path.extname(file) === '.mdx' || path.extname(file) === '.md');
+}
